10127 admin rendering for FAM and TPM partners
